Issue/420 fetch announcements (#421)
* Add checks for `context_code` key to announcements

Announcements come in through the `Canvas` object and do not include
a course ID directly. Since announcements are DiscussionTopics, we
can parse the ID from the `context_code` param.

Currently, the Canvas API only supports course-level context codes,
so this works for now, but will break as more contexts are added.

https://canvas.instructure.com/doc/api/announcements.html#method.announcements_api.index

* Update tests for announcements course ID parse

* ignore venv directory

* moved venv with other env/ directories

* more thorough checks for context_codes

* Accept a list of course IDs in get_announcements

Canvas returns announcements from the `canvas` object. The user has to
define which course or courses to pull from. Adding `course_list` as a
required parameter prevents failed calls by exposing this required
parameter to the user.

The list is processed to the `context_codes` format before being sent
in the request.

* update tests

* fix black formatting, added test coverage

* Add checks for `context_code` key to announcements

Announcements come in through the `Canvas` object and do not include
a course ID directly. Since announcements are DiscussionTopics, we
can parse the ID from the `context_code` param.

Currently, the Canvas API only supports course-level context codes,
so this works for now, but will break as more contexts are added.

https://canvas.instructure.com/doc/api/announcements.html#method.announcements_api.index

* Update tests for announcements course ID parse

* ignore venv directory

* moved venv with other env/ directories

* more thorough checks for context_codes

* Accept a list of course IDs in get_announcements

Canvas returns announcements from the `canvas` object. The user has to
define which course or courses to pull from. Adding `course_list` as a
required parameter prevents failed calls by exposing this required
parameter to the user.

The list is processed to the `context_codes` format before being sent
in the request.

* update tests

* fix black formatting, added test coverage

* Refactor for more flexible arguments

See https://github.com/ucfopen/canvasapi/pull/421#discussion_r499841262

The requester can handle `Course` objects or plain strings. Handling both
makes it easier on the user because they can reuse a course object they may
already have. Canvas just looks for the `context_code` string.

The `context_codes` param is required by the API. The `get_announcements` function
requires a list of either strings or `Course` objects to request. Formatting
is done by the library before sending to the `Announcements` endpoint.

* Added tests

Wrote tests specifically to accept `Course` objects as items in the
`context_codes` list used to get announcements.

* add newline at end

* Simpler context code checking, updated tests

Used `startswith()` to check for the type of code rather than `re`.
Finalize logic for taking in `<Course>` objects or course IDs for fetching
announcements.

* remove merge conflict

* black formatting

Co-authored-by: Matthew Emond <me@ucf.edu>